+++ apx_core/src/json_signatures/verify.rs
@@ -448,7 +448,7 @@ mod tests {
         let object = json!({
             "@context": [
                 "https://www.w3.org/ns/activitystreams",
-                "https://w3id.org/security/data-integrity/v1"
+                "https://w3id.org/security/data-integrity/v2"
             ],
             "id": "https://server.example/activities/1",
             "type": "Create",
@@ -478,7 +478,7 @@ mod tests {
         let expected_result = json!({
             "@context": [
                 "https://www.w3.org/ns/activitystreams",
-                "https://w3id.org/security/data-integrity/v1"
+                "https://w3id.org/security/data-integrity/v2"
             ],
             "id": "https://server.example/activities/1",
             "type": "Create",
@@ -497,13 +497,13 @@ mod tests {
             "proof": {
                 "@context": [
                     "https://www.w3.org/ns/activitystreams",
-                    "https://w3id.org/security/data-integrity/v1"
+                    "https://w3id.org/security/data-integrity/v2"
                 ],
                 "type": "DataIntegrityProof",
                 "cryptosuite": "eddsa-jcs-2022",
                 "verificationMethod": "https://server.example/users/alice#ed25519-key",
                 "proofPurpose": "assertionMethod",
-                "proofValue": "zLaewdp4H9kqtwyrLatK4cjY5oRHwVcw4gibPSUDYDMhi4M49v8pcYk3ZB6D69dNpAPbUmY8ocuJ3m9KhKJEEg7z",
+                "proofValue": "z42ffGu6AUKPCFcFPiabmUvnGLPJzC7e4DGWC52NUasSSH37UMa9c58tdgVszUcZfytxa4fQ5TYHaJENCxUDe9SdL",
                 "created": "2023-02-24T23:36:38Z"
             }
         });
@@ -525,9 +525,9 @@ mod tests {
         assert_eq!(result.is_ok(), true);
     }
 
-    /// https://w3c.github.io/vc-di-eddsa/#representation-eddsa-jcs-2022
     #[test]
     fn test_create_and_verify_eddsa_signature_vc_di_eddsa_test_vector() {
+        // https://w3c.github.io/vc-di-eddsa/#representation-eddsa-jcs-2022
         let secret_key_multibase = "z3u2en7t5LR2WtQH5PfFqMqwVHBeXouLzo6haApm8XHqvjxq";
         let secret_key = ed25519_secret_key_from_multikey(secret_key_multibase).unwrap();
         let key_id = "did:key:z6MkrJVnaZkeFzdQyMZu1cgjg7k1pZZ6pvBQ7XJPt4swbTQ2#z6MkrJVnaZkeFzdQyMZu1cgjg7k1pZZ6pvBQ7XJPt4swbTQ2";