+++ mitra_activitypub/src/builders/note.rs
@@ -581,7 +581,7 @@ mod tests {
             "@context": [
                 "https://www.w3.org/ns/activitystreams",
                 "https://w3id.org/security/v1",
-                "https://w3id.org/security/data-integrity/v1",
+                "https://w3id.org/security/data-integrity/v2",
                 {
                     "Hashtag": "as:Hashtag",
                     "sensitive": "as:sensitive",
@@ -842,7 +842,7 @@ mod tests {
             "@context": [
                 "https://www.w3.org/ns/activitystreams",
                 "https://w3id.org/security/v1",
-                "https://w3id.org/security/data-integrity/v1",
+                "https://w3id.org/security/data-integrity/v2",
                 {
                     "Hashtag": "as:Hashtag",
                     "sensitive": "as:sensitive",
@@ -910,7 +910,7 @@ mod tests {
             "@context": [
                 "https://www.w3.org/ns/activitystreams",
                 "https://w3id.org/security/v1",
-                "https://w3id.org/security/data-integrity/v1",
+                "https://w3id.org/security/data-integrity/v2",
                 {
                     "Hashtag": "as:Hashtag",
                     "sensitive": "as:sensitive",