+++ mitra_activitypub/src/contexts.rs
@@ -6,7 +6,7 @@ pub use apx_sdk::constants::{
 
 pub const W3C_CID_CONTEXT: &str = "https://www.w3.org/ns/cid/v1";
 pub const W3ID_SECURITY_CONTEXT: &str = "https://w3id.org/security/v1";
-pub const W3ID_DATA_INTEGRITY_CONTEXT: &str = "https://w3id.org/security/data-integrity/v1";
+pub const W3ID_DATA_INTEGRITY_CONTEXT: &str = "https://w3id.org/security/data-integrity/v2";
 pub const W3ID_VALUEFLOWS_CONTEXT: &str = "https://w3id.org/valueflows/ont/vf#";
 pub const SCHEMA_ORG_CONTEXT: &str = "http://schema.org/";
 pub const MASTODON_CONTEXT: &str = "http://joinmastodon.org/ns#";