+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Changed data integrity context to `https://w3id.org/security/data-integrity/v2`.
+
 ## [4.11.0] - 2025-10-09
 
 ### Added